Changed and fixed the internal mechanism of showing and hiding pages.
► Merged "_init_widget_tree" of "main.py" to the constrcutor.
► Moved some code from "_post_update" back to "_constant_updates" in "main.py".
Realized that 'pos' is not bound to the window at all when resizing. I thought
that when you bound some of the object's attributes to the root widget. It will
become a part of its life cycle and update automatically when resizing the
window. Apparently it does not, and there is no such. This is the result of not
reading the documentation enough lol. Back to work..
► Fixed toggle page mechanism of "FrontWidget", "AppInfo",
"IngredientsDataPanel", and "AddIngredientButton" kv rules.
► Touched up code and cleaned comments.