Fixed reference issue popping out of nowhere when trying out the new popup.
► The issue displays itself after doing the following:
    ♦ After opening the app.
    ♦ Go to menu,
    ♦ then click "Manage Ingredients",
    ♦ then press "Add Custom Ingredient".
    ♦ Fill out all the input entries with "123" without quotes.
    ♦ Set all the effects to "Cure Disease" for simplisity.
    ♦ Go back to Value entry and add a letter a there. Should be "123a" now.
    ♦ Click "Add Ingredient" button. Error should now pop up.
    ♦ Click the "Cancel Button".
    ♦ Then select and press any ingredient to auto search. i. e. "Bee".
    ♦ You should be back to the front page with the result of the ingredient.
    ♦ Go back to the "Add Ingredient" page.
    ♦ Remove the letter a from Value entry. Should now be "123".
    ♦ Finally press "Add Ingredient" button.

Dev Note.
It should be noted that when I tried the steps above after before fixing the
reference issue. The "Adding Failed" label no longer pops up for me. Previously
before tampering with anything using git. My program crashes resulting in a
NoneType value error. I dug around using pdb and found out that the issue lies
within the weakref I used to add the widget reference of "error_title" at
"panel_add_ingredient.py:line167" to it's parent's dictionary of id. Therefore
when the program tries to remove the previously created widgets when the check
is finally finished. The program crashes because it tries to remove something
that hasn't been referenced properly. Or sometimes it just doesn't add the
labels to the designated parent resulting in the warning label not showing up
at the "Add Ingredient" panel even though some values may be missing or
incorrect. Use git diff to see the changes I have made in more detail.

► Replaced code that manually adds the warning labels of "Add Ingredient"
panel's id to its ids dictionary.
► Changed code checking for the warling label's id in the ids to instead
manually inspect the children of the grid of the panel.
► Removed "CustomLabel" rule template of "panel_application_info.kv".