Finish method to append ingredient entry to the database.
Confirming custom ingredient entry now appends it to the database. Although
the ingredients inside the "Manage Ingredients" panel and searching it wont
show up yet. But I have confirmed that it does stay inside the database.
Therefor even tho you can't search it in the session you added it neither can
you see it in the "Manage Ingredients" panel. It shows up at both once you
restart the app. Although I'll work on explicitly updating the app for that.
One more thing I forgot to mention is that the custom ingredient you add to the
database also is being read by the spinners of the "Add Ingredient" panel. This
concludes this commit.

► "add_ingredient_panel.py > reset_entries" now also resets the panel's error
labels.
► Changed bits of reference.
► Renamed "add_ingredient_panel.py > AddIngredientPanel> add_ingredient" to
"prep_entry".
► Created new method "add_ingredient_panel.py > AddIngredientPanel >
append_entry" that appends the ingredient entry to the database.
► Renamed and finished method "db_api.py > Database > add_ingredient" to
"append_ingredient".